More auto-style updates

+++ M3DSoftware/GUI/Views/Printer_View/Colors.cs
@@ -5,22 +5,22 @@ namespace M3D.GUI.Views.Printer_View
 {
   internal static class Colors
   {
-    public static readonly Color4 black = new Color4((byte) 52, (byte) 52, (byte) 52, byte.MaxValue);
-    public static readonly Color4 silver = new Color4((byte) 185, (byte) 185, (byte) 185, byte.MaxValue);
-    public static readonly Color4 orange = new Color4(byte.MaxValue, (byte) 70, (byte) 0, byte.MaxValue);
-    public static readonly Color4 green = new Color4((byte) 66, (byte) 187, (byte) 49, byte.MaxValue);
-    public static readonly Color4 gray = new Color4((byte) 102, (byte) 102, (byte) 102, byte.MaxValue);
-    public static readonly Color4 darkBlue = new Color4((byte) 1, (byte) 0, (byte) 138, byte.MaxValue);
-    public static readonly Color4 grape = new Color4((byte) 93, (byte) 24, (byte) 106, byte.MaxValue);
-    public static readonly Color4 lightBlue = new Color4((byte) 98, (byte) 181, (byte) 233, byte.MaxValue);
-    public static readonly Color4 lightGreen = new Color4((byte) 144, (byte) 238, (byte) 144, byte.MaxValue);
-    public static readonly Color4 natural = new Color4((byte) 0, (byte) 159, (byte) 236, byte.MaxValue);
-    public static readonly Color4 neonOrange = new Color4(byte.MaxValue, (byte) 104, (byte) 1, byte.MaxValue);
-    public static readonly Color4 red = new Color4((byte) 196, (byte) 4, (byte) 1, byte.MaxValue);
+    public static readonly Color4 black = new Color4(52, 52, 52, byte.MaxValue);
+    public static readonly Color4 silver = new Color4(185, 185, 185, byte.MaxValue);
+    public static readonly Color4 orange = new Color4(byte.MaxValue, 70, 0, byte.MaxValue);
+    public static readonly Color4 green = new Color4(66, 187, 49, byte.MaxValue);
+    public static readonly Color4 gray = new Color4(102, 102, 102, byte.MaxValue);
+    public static readonly Color4 darkBlue = new Color4(1, 0, 138, byte.MaxValue);
+    public static readonly Color4 grape = new Color4(93, 24, 106, byte.MaxValue);
+    public static readonly Color4 lightBlue = new Color4(98, 181, 233, byte.MaxValue);
+    public static readonly Color4 lightGreen = new Color4(144, 238, 144, byte.MaxValue);
+    public static readonly Color4 natural = new Color4(0, 159, 236, byte.MaxValue);
+    public static readonly Color4 neonOrange = new Color4(byte.MaxValue, 104, 1, byte.MaxValue);
+    public static readonly Color4 red = new Color4(196, 4, 1, byte.MaxValue);
     public static readonly Color4 white = new Color4(byte.MaxValue, byte.MaxValue, byte.MaxValue, byte.MaxValue);
-    public static readonly Color4 brown = new Color4((byte) 101, (byte) 66, (byte) 34, byte.MaxValue);
-    public static readonly Color4 yellow = new Color4(byte.MaxValue, (byte) 215, (byte) 0, byte.MaxValue);
-    public static readonly Color4 clear = new Color4((byte) 105, (byte) 100, (byte) 110, (byte) 75);
+    public static readonly Color4 brown = new Color4(101, 66, 34, byte.MaxValue);
+    public static readonly Color4 yellow = new Color4(byte.MaxValue, 215, 0, byte.MaxValue);
+    public static readonly Color4 clear = new Color4(105, 100, 110, 75);
 
     public static Color4 GetColorFromSN(string serialNumber)
     {
