+++ M3DSpooling/M3D/Spooling/FirstRunUpdates/FirstRunUpdateResetOffsets.cs
@@ -12,7 +12,7 @@ namespace M3D.Spooling.FirstRunUpdates
     {
       try
       {
-        var uint32 = (int) BitConverter.ToUInt32(eeprom, (int) printerProfile.EEPROMConstants.GetEepromInfo("FirmwareVersion").EepromAddr);
+        var uint32 = (int) BitConverter.ToUInt32(eeprom, printerProfile.EEPROMConstants.GetEepromInfo("FirmwareVersion").EepromAddr);
         var eepromAddr1 = (int) printerProfile.EEPROMConstants.GetEepromInfo("ZCalibrationBLO").EepromAddr;
         var eepromAddr2 = (int) printerProfile.EEPROMConstants.GetEepromInfo("ZCalibrationBRO").EepromAddr;
         var eepromAddr3 = (int) printerProfile.EEPROMConstants.GetEepromInfo("ZCalibrationFLO").EepromAddr;
@@ -27,7 +27,7 @@ namespace M3D.Spooling.FirstRunUpdates
         var num2 = 2015080402;
         if ((uint) uint32 < (uint) num2)
         {
-          if ((double) Math.Abs(single1) > 1.40129846432482E-45 || (double) Math.Abs(single2) > 1.40129846432482E-45 || ((double) Math.Abs(single3) > 1.40129846432482E-45 || (double) Math.Abs(single4) > 1.40129846432482E-45) || (double) Math.Abs(single5) > 1.40129846432482E-45)
+          if (Math.Abs(single1) > 1.40129846432482E-45 || Math.Abs(single2) > 1.40129846432482E-45 || (Math.Abs(single3) > 1.40129846432482E-45 || Math.Abs(single4) > 1.40129846432482E-45) || Math.Abs(single5) > 1.40129846432482E-45)
           {
             try
             {
